+++ src/config/routetree.php
@@ -17,6 +17,20 @@ return [
 
     /*
     |--------------------------------------------------------------------------
+    | Locales with no automatic path prefix.
+    |--------------------------------------------------------------------------
+    |
+    | This setting will disable prefixing of path with locale for this language.
+    | (e.g. setting this to ['de'], if ['en','de'] is configured as locales above
+    | will result in the paths '/kontakt' for 'de' and 'en/contact' for 'en'.
+    |
+    | Attention: This can result in ambiguous paths!
+    |
+    */
+    'no_prefix_locales' => [],
+
+    /*
+    |--------------------------------------------------------------------------
     | Create absolute paths instead of relative paths by default?
     |--------------------------------------------------------------------------
     |
@@ -79,19 +93,35 @@ return [
 
     /*
     |--------------------------------------------------------------------------
-    | Sitemap Settings
+    | XML-Sitemap Settings
     |--------------------------------------------------------------------------
     |
-    | Here you may configure the settings for generation
-    | of a sitemap XML file via the artisan command 'routetree:generate-sitemap'.
+    | Here you may configure the settings for generation of an XML-sitemap.
     |
     */
     'sitemap' => [
 
         /*
-         * Name of the output file (relative to laravel root).
+         * Name of the output file (relative to laravel root)
+         * of the artisan command 'routetree:generate-sitemap'.
          */
         'output_file' => 'public/sitemap.xml',
+
+        /*
+         * Settings related to a route for delivering a
+         * dynamic sitemap.
+         */
+        'route' => [
+
+            // Enable the sitemap route.
+            // RouteTree will automatically
+            // register the route, if enabled.
+            'enabled' => false,
+
+            // Path to the dynamic XML-sitemap
+            // (e.g. 'sitemap.xml').
+            'path' => 'sitemap.xml',
+        ],
 
         /*
          * Base URL of generated urls.