+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release Notes
 
+## [v2.3.0 (2020-08-21)](https://github.com/webflorist/routetree/compare/v2.2.1...v2.3.0)
+### Added
+- Added config `sitemap.route` to enable a route delivering a dynamic XML-sitemap.
+- Added config `no_prefix_locales` to disable locale-prefix for specific languages.
+
 ## [v2.2.1 (2020-03-04)](https://github.com/webflorist/routetree/compare/v2.2.0...v2.2.1)
 ### Added
 - Add method `omitActionNameFromRouteName` to RouteAction to circumvent problems with legacy route names.